Implement Gzip compression; Auto-reboot on completion; Restart into the client OS
Rebooting into the client has been done using a simple pre-existing
commandline tool called efibootmgr. Given that we are using a Debian
GNU/Linux system as the management OS on non-ancient hardware, this
should work the target computer.

Using gzip serves two purposes: the first of which is that it is a
standard algorithm which gives broad architectural, operating system
and language. Secondly, it is used both during network transfer and
on disk whic means that we can avoid the complicated dependency on
file systems with transparent encryption like btrfs or ZFS. They can
still be used, of course, but it is no longer a hard requirement which
makes it easier to deploy on university networks.